Integrate generic reuse regressions into existing input tests
