+++ tests/test_input_model.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 import importlib
 import json
 import os
+import subprocess
 import sys
 import types
 from argparse import Namespace
@@ -16,11 +17,14 @@ from typing import TYPE_CHECKING, Any
 
 import pytest
 from pydantic import TypeAdapter, ValidationError
+from pydantic.version import VERSION as PYDANTIC_VERSION
 
 from datamodel_code_generator import DataModelType, GenerateConfig, InputFileType, arguments
 from datamodel_code_generator import __main__ as main_module
 from datamodel_code_generator.__main__ import Exit
+from datamodel_code_generator.enums import InputModelRefStrategy
 from datamodel_code_generator.format import Formatter
+from datamodel_code_generator.input_model import Error as InputModelError
 from datamodel_code_generator.input_model import load_model_schema
 from tests.conftest import (
     BUILTIN_FORMATTER_VALUE,
@@ -32,7 +36,7 @@ from tests.conftest import (
 from tests.data.python.input_model import inherited_overrides, union_annotations
 from tests.data.python.input_model.inherited_override_runtime import CASES as INHERITED_OVERRIDE_CASES
 from tests.data.python.input_model.union_runtime import CASES, VALUES, describe
-from tests.main.conftest import _generated_model, run_generate_and_assert, run_main_and_assert, run_main_with_args
+from tests.main.conftest import DATA_PATH, EXPECTED_MAIN_PATH, _generated_model, run_generate_and_assert, run_main_and_assert, run_main_with_args
 
 if TYPE_CHECKING:
     from collections.abc import Iterator, Sequence
@@ -2742,3 +2746,160 @@ def test_input_model_dynamic_nested_exports(
         assert nested_dynamic_exports.calls == (
             ["__path__", case] if strategy != "regenerate-all" and case not in {"Live", "Lazy", "Proxy"} else []
         )
+
+
+INPUT_GENERIC_FIXTURES = DATA_PATH / "python/input_model"
+
+
+INPUT_GENERIC_EXPECTED = EXPECTED_MAIN_PATH / "input_model/generic_reuse"
+
+
+INPUT_GENERIC_CASES = json.loads((INPUT_GENERIC_FIXTURES / "generic_reuse_cases.json").read_text())
+
+
+@pytest.mark.parametrize(
+    ("record", "strategy"),
+    [
+        pytest.param(record, strategy, id=f"{strategy}-{record['name']}")
+        for record in INPUT_GENERIC_CASES
+        for strategy in record.get("strategies", ["regenerate-all", "reuse-all", "reuse-foreign"])
+    ],
+)
+@pytest.mark.parametrize("entrypoint", ["cli", "api"])
+@pytest.mark.parametrize("formatter", ["builtin", "external"])
+def test_python_generic_reuse(record: dict, strategy: str, entrypoint: str, formatter: str, tmp_path: Path) -> None:
+    """Preserve original validators and specialization identity without changing regeneration."""
+    paths = ["tests.data.python.input_model." + source for source in record["sources"]]
+    (tmp_path / "pyproject.toml").write_text((INPUT_GENERIC_FIXTURES / "collision_settings/pyproject.toml").read_text())
+    formatters = ["builtin"] if formatter == "builtin" else ["black", "isort"]
+    output = tmp_path / "output.py"
+    if entrypoint == "cli":
+        args = [argument for path in paths for argument in ("--input-model", path)]
+        run_main_with_args([
+            *args,
+            "--output",
+            str(output),
+            "--disable-timestamp",
+            "--input-model-ref-strategy",
+            strategy,
+            "--formatters",
+            *formatters,
+        ])
+    else:
+        schema = load_model_schema(
+            paths, InputFileType.JsonSchema, InputModelRefStrategy(strategy), DataModelType.PydanticV2BaseModel
+        )
+        schema = json.loads(json.dumps(schema))
+        config = GenerateConfig(
+            input_file_type=InputFileType.JsonSchema,
+            disable_timestamp=True,
+            input_filename="<stdin>",
+            output=output,
+            settings_path=tmp_path,
+            formatters=[Formatter(value) for value in formatters],
+        )
+        generate(schema, config=config)
+    expected_strategy = "regenerate-all" if strategy == "regenerate-all" else "reuse-all"
+    stem = record["name"] + "_" + expected_strategy
+    source_expected = INPUT_GENERIC_EXPECTED / formatter if record.get("formatter_goldens") else INPUT_GENERIC_EXPECTED
+    if record.get("pydantic20_goldens") and PYDANTIC_VERSION.split(".")[:2] == ["2", "0"]:
+        source_expected = INPUT_GENERIC_EXPECTED / "pydantic20" / formatter
+    assert_output(output.read_text(), source_expected / (stem + ".py"))
+    runtime = subprocess.run(
+        [sys.executable, str(INPUT_GENERIC_FIXTURES / "generic_reuse_runtime.py"), str(output), json.dumps(record)],
+        capture_output=True,
+        text=True,
+        check=True,
+    )
+    assert_output(runtime.stdout, INPUT_GENERIC_EXPECTED / (stem + ".txt"))
+
+
+@pytest.mark.parametrize("entrypoint", ["cli", "api"])
+@pytest.mark.parametrize("formatter", ["builtin", "external"])
+def test_python_generic_reuse_type_binding(entrypoint: str, formatter: str, tmp_path: Path) -> None:
+    """Resolve a transported generic definition from a preserved field expression."""
+    fixture = DATA_PATH / "jsonschema/generic_reuse_type.json"
+    output = tmp_path / "output.py"
+    (tmp_path / "pyproject.toml").write_text((INPUT_GENERIC_FIXTURES / "collision_settings/pyproject.toml").read_text())
+    formatters = ["builtin"] if formatter == "builtin" else ["black", "isort"]
+    if entrypoint == "cli":
+        run_main_with_args([
+            "--input",
+            str(fixture),
+            "--output",
+            str(output),
+            "--input-file-type",
+            "jsonschema",
+            "--disable-timestamp",
+            "--formatters",
+            *formatters,
+        ])
+    else:
+        generate(
+            fixture,
+            config=GenerateConfig(
+                input_file_type=InputFileType.JsonSchema,
+                output=output,
+                disable_timestamp=True,
+                settings_path=tmp_path,
+                formatters=[Formatter(value) for value in formatters],
+            ),
+        )
+    assert_output(output.read_text(), INPUT_GENERIC_EXPECTED / "preserved.py")
+    runtime = subprocess.run(
+        [
+            sys.executable,
+            str(INPUT_GENERIC_FIXTURES / "generic_reuse_runtime.py"),
+            str(output),
+            json.dumps(INPUT_GENERIC_CASES[0]),
+        ],
+        capture_output=True,
+        text=True,
+        check=True,
+    )
+    assert_output(runtime.stdout, INPUT_GENERIC_EXPECTED / "int_reuse-all.txt")
+
+
+@pytest.mark.parametrize("case", ["name", "family", "index_type", "index_negative"])
+def test_python_native_field_invalid_path(case: str, tmp_path: Path) -> None:
+    """Reject malformed external native paths before generating executable annotations."""
+    record = json.loads((DATA_PATH / "payloads/generic_native_fields" / (case + ".json")).read_text())
+    with pytest.raises((TypeError, ValueError), match=record["message"]):
+        generate(
+            record["schema"],
+            config=GenerateConfig(input_file_type=InputFileType.JsonSchema, output=tmp_path / "output.py"),
+        )
+
+
+@pytest.mark.parametrize("entrypoint", ["cli", "api"])
+@pytest.mark.parametrize("strategy", ["reuse-all", "reuse-foreign"])
+@pytest.mark.parametrize("formatter", ["builtin", "external"])
+def test_python_inline_future_generic_diagnostic(
+    entrypoint: str, strategy: str, formatter: str, capsys: pytest.CaptureFixture[str], tmp_path: Path
+) -> None:
+    """Diagnose only metadata which has neither a native path nor a stable export."""
+    fixture = json.loads((INPUT_GENERIC_FIXTURES / "generic_reuse_diagnostic.json").read_text())
+    if entrypoint == "cli":
+        run_main_with_args(
+            [
+                "--input-model",
+                fixture["source"],
+                "--input-model-ref-strategy",
+                strategy,
+                "--output",
+                str(tmp_path / "output.py"),
+                "--formatters",
+                *(["builtin"] if formatter == "builtin" else ["black", "isort"]),
+            ],
+            expected_exit=Exit.ERROR,
+            capsys=capsys,
+            expected_stderr_contains=fixture["message"],
+        )
+    else:
+        with pytest.raises(InputModelError, match=fixture["message"]):
+            load_model_schema(
+                [fixture["source"]],
+                InputFileType.JsonSchema,
+                InputModelRefStrategy(strategy),
+                DataModelType.PydanticV2BaseModel,
+            )
